Lots of random changes which include: - Started work on Bounty system. - CombatLog & CombatTimer have been fully implemented. - ControlPoints system added. - New Donator effects have been added. - Donator commands have been added. - Economy has been fully implemented. - Hohengroth Event & Event framework has been started. - Lootchests have been fully implemented. - New Spells GUI has been added. - Focus for Squads has been implemented - Staff commands have been added. - StaffMode and staff security has been implemented. - Tickets system has been implemented. - Hit indicators have been added.

+++ src/com/medievallords/carbyne/crates/CrateManager.java
@@ -111,7 +111,7 @@ public class CrateManager {
                             List<String> enchantments = rewardsSection.getStringList(rewardId + ".Enchantments");
                             List<String> commands = rewardsSection.getStringList(rewardId + ".Commands");
                             boolean displayItemOnly = rewardsSection.getBoolean(rewardId + ".DisplayItem");
-                            int chance = rewardsSection.getInt(rewardId + ".Chance");
+                            double chance = rewardsSection.getDouble(rewardId + ".Chance");
 
                             if (displayName != null) {
                                 if (displayName.contains("randomgear")) {
